Refactor color_rgb_clamp to use float instead of double to fix compiler float conversion warnings

+++ src/bacnet/basic/sys/color_rgb.c
@@ -14,18 +14,18 @@
 #include "bacnet/basic/sys/color_rgb.h"
 
 /**
- * @brief Clamp a double precision value between two limits
+ * @brief Clamp a single precision value between two limits
  * @param d - value to be clamped
  * @param min - minimum value to clamp within
  * @param max - maximum value to clamp within
  * @return value clamped between min and max inclusive
  */
-double color_rgb_clamp(double d, double min, double max)
+float color_rgb_clamp(float d, float min, float max)
 {
     if (isnan(d)) {
         return min;
     } else {
-        const double t = d < min ? min : d;
+        const float t = d < min ? min : d;
         return t > max ? max : t;
     }
 }
@@ -587,19 +587,19 @@ void color_rgb_from_temperature(
         red = 255.0;
     } else {
         red = (float)(temperature_kelvin - 60);
-        red = 329.698727446 * pow(red, -0.1332047592);
-        red = color_rgb_clamp(red, 0.0, 255.0);
+        red = (float)(329.698727446 * pow(red, -0.1332047592));
+        red = color_rgb_clamp(red, 0.0f, 255.0f);
     }
     /* Calculate Green */
     if (temperature_kelvin <= 66) {
         /* Green values below 6600 K */
         green = (float)temperature_kelvin;
-        green = 99.4708025861 * log(green) - 161.1195681661;
+        green = (float)(99.4708025861 * log(green) - 161.1195681661);
     } else {
         green = (float)(temperature_kelvin - 60);
-        green = 288.1221695283 * pow(green, -0.0755148492);
+        green = (float)(288.1221695283 * pow(green, -0.0755148492));
     }
-    green = color_rgb_clamp(green, 0.0, 255.0);
+    green = color_rgb_clamp(green, 0.0f, 255.0f);
     /* Calculate Blue */
     if (temperature_kelvin >= 66) {
         /* Blue values above 6600 K */
@@ -609,8 +609,8 @@ void color_rgb_from_temperature(
         blue = 0.0;
     } else {
         blue = (float)(temperature_kelvin - 10);
-        blue = 138.5177312231 * log(blue) - 305.0447927307;
-        blue = color_rgb_clamp(blue, 0, 255);
+        blue = (float)(138.5177312231 * log(blue) - 305.0447927307);
+        blue = color_rgb_clamp(blue, 0.0f, 255.0f);
     }
     if (r) {
         *r = (uint8_t)red;
